Migrate ChunkState and its tests under where UserArtifactsManager is at
